Ai_order/payment_hub/requirements.txt
flask
python-dotenv
requests
psycopg2-binary
